# backend/YanCarz/swagger.json
{
    "openapi":  "3.0.4",
    "info":  {
                 "title":  "YanCarz.API",
                 "version":  "1.0"
             },
    "paths":  {
                  "/api/portal/Agency":  {
                                             "get":  {
                                                         "tags":  [
                                                                      "Agency"
                                                                  ],
                                                         "responses":  {
                                                                           "200":  {
                                                                                       "description":  "OK",
                                                                                       "content":  {
                                                                                                       "text/plain":  {
                                                                                                                          "schema":  {
                                                                                                                                         "type":  "array",
                                                                                                                                         "items":  {
                                                                                                                                                       "$ref":  "#/components/schemas/AgencyDto"
                                                                                                                                                   }
                                                                                                                                     }
                                                                                                                      },
                                                                                                       "application/json":  {
                                                                                                                                "schema":  {
                                                                                                                                               "type":  "array",
                                                                                                                                               "items":  {
                                                                                                                                                             "$ref":  "#/components/schemas/AgencyDto"
                                                                                                                                                         }
                                                                                                                                           }
                                                                                                                            },
                                                                                                       "text/json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "type":  "array",
                                                                                                                                        "items":  {
                                                                                                                                                      "$ref":  "#/components/schemas/AgencyDto"
                                                                                                                                                  }
                                                                                                                                    }
                                                                                                                     }
                                                                                                   }
                                                                                   }
                                                                       }
                                                     },
                                             "post":  {
                                                          "tags":  [
                                                                       "Agency"
                                                                   ],
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/AgencyCreateDto"
                                                                                                                                  }
                                                                                                                   },
                                                                                              "text/json":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/AgencyCreateDto"
                                                                                                                           }
                                                                                                            },
                                                                                              "application/*+json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/AgencyCreateDto"
                                                                                                                                    }
                                                                                                                     }
                                                                                          }
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK"
                                                                                    }
                                                                        }
                                                      }
                                         },
                  "/api/portal/Agency/{id}":  {
                                                  "get":  {
                                                              "tags":  [
                                                                           "Agency"
                                                                       ],
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "id",
                                                                                     "in":  "path",
                                                                                     "required":  true,
                                                                                     "schema":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "uuid"
                                                                                                }
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK",
                                                                                            "content":  {
                                                                                                            "text/plain":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/AgencyDto"
                                                                                                                                          }
                                                                                                                           },
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "$ref":  "#/components/schemas/AgencyDto"
                                                                                                                                                }
                                                                                                                                 },
                                                                                                            "text/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/AgencyDto"
                                                                                                                                         }
                                                                                                                          }
                                                                                                        }
                                                                                        }
                                                                            }
                                                          }
                                              },
                  "/api/agency/Agency/{id}":  {
                                                  "get":  {
                                                              "tags":  [
                                                                           "Agency"
                                                                       ],
                                                              "parameters":  [
                                                                                 {
                                                                                     "name":  "id",
                                                                                     "in":  "path",
                                                                                     "required":  true,
                                                                                     "schema":  {
                                                                                                    "type":  "string",
                                                                                                    "format":  "uuid"
                                                                                                }
                                                                                 }
                                                                             ],
                                                              "responses":  {
                                                                                "200":  {
                                                                                            "description":  "OK",
                                                                                            "content":  {
                                                                                                            "text/plain":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/AgencyDto"
                                                                                                                                          }
                                                                                                                           },
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "$ref":  "#/components/schemas/AgencyDto"
                                                                                                                                                }
                                                                                                                                 },
                                                                                                            "text/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/AgencyDto"
                                                                                                                                         }
                                                                                                                          }
                                                                                                        }
                                                                                        }
                                                                            }
                                                          }
                                              },
                  "/api/agency/Agency":  {
                                             "post":  {
                                                          "tags":  [
                                                                       "Agency"
                                                                   ],
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/AgencyCreateDto"
                                                                                                                                  }
                                                                                                                   },
                                                                                              "text/json":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/AgencyCreateDto"
                                                                                                                           }
                                                                                                            },
                                                                                              "application/*+json":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/AgencyCreateDto"
                                                                                                                                    }
                                                                                                                     }
                                                                                          }
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK"
                                                                                    }
                                                                        }
                                                      }
                                         },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "AgencyCreateDto":  {
                                                               "type":  "object",
                                                               "properties":  {
                                                                                  "name":  {
                                                                                               "type":  "string",
                                                                                               "nullable":  true
                                                                                           },
                                                                                  "eMail":  {
                                                                                                "type":  "string",
                                                                                                "nullable":  true
                                                                                            },
                                                                                  "nbrPhone":  {
                                                                                                   "type":  "string",
                                                                                                   "nullable":  true
                                                                                               },
                                                                                  "lastName":  {
                                                                                                   "type":  "string",
                                                                                                   "nullable":  true
                                                                                               },
                                                                                  "firstMame":  {
                                                                                                    "type":  "string",
                                                                                                    "nullable":  true
                                                                                                },
                                                                                  "address":  {
                                                                                                  "type":  "string",
                                                                                                  "nullable":  true
                                                                                              },
                                                                                  "idCity":  {
                                                                                                 "type":  "string",
                                                                                                 "format":  "uuid"
                                                                                             }
                                                                              },
                                                               "additionalProperties":  false
                                                           },
                                       "AgencyDto":  {
                                                         "type":  "object",
                                                         "properties":  {
                                                                            "id":  {
                                                                                       "type":  "string",
                                                                                       "format":  "uuid"
                                                                                   },
                                                                            "name":  {
                                                                                         "type":  "string",
                                                                                         "nullable":  true
                                                                                     },
                                                                            "eMail":  {
                                                                                          "type":  "string",
                                                                                          "nullable":  true
                                                                                      },
                                                                            "nbrPhone":  {
                                                                                             "type":  "string",
                                                                                             "nullable":  true
                                                                                         },
                                                                            "address":  {
                                                                                            "type":  "string",
                                                                                            "nullable":  true
                                                                                        }
                                                                        },
                                                         "additionalProperties":  false
                                                     },
...
                                   }
                   },
...
}